Move UserAvatarHeader back to church app
Step one in refactoring UserAvatarHeader. Another PR will follow to
update the GET_USER_PROFILE related files.
