+++ packages/apolloschurchapp/src/tabs/connect/Connect.js
@@ -3,14 +3,12 @@ import { ScrollView } from 'react-native';
 import { SafeAreaView } from 'react-navigation';
 import PropTypes from 'prop-types';
 
-import {
-  HorizontalLikedContentFeedConnected,
-  UserAvatarHeader,
-} from '@apollosproject/ui-connected';
+import { HorizontalLikedContentFeedConnected } from '@apollosproject/ui-connected';
 import { BackgroundView } from '@apollosproject/ui-kit';
 
 import ActionTable from './ActionTable';
 import ActionBar from './ActionBar';
+import UserAvatarHeader from './UserAvatarHeader';
 
 class Connect extends PureComponent {
   static navigationOptions = () => ({
