+++ packages/apolloschurchapp/src/tabs/connect/UserAvatarHeader/getUserProfile.js
