+++ packages/apolloschurchapp/src/tabs/connect/UserAvatarHeader/index.js
